pipeline_project/requirements.txt
Django>=4.2
djangorestframework>=3.14
opencv-python>=4.7
numpy>=1.25
Pillow>=10.0
drf-spectacular>=0.27
pdf2image>=1.16
Django>=4.2
djangorestframework>=3.14
opencv-python>=4.7
numpy>=1.25
Pillow>=10.0
drf-spectacular>=0.27
Django
djangorestframework
opencv-python
numpy
Pillow